add logging about sending diagnostics to client
Summary:
We have a problem where sometimes (unsure how common), diagnostics aren't getting sent from server to client after recheck.

This diff adds a single line per recheck to the server log so that, in cases where the server declined to send the message, we'll be able to read why the server declined.

Reviewed By: 2BitSalute

Differential Revision: D16389931

fbshipit-source-id: 43dd41deb4909a063c60650ae78ed4771199fbc2