server rage tells about client
Summary:
Here is more logging to help track down the cause for why sometimes diagnostic messages aren't getting pushed to the client.

The effect of this diff is that, whenever you click the Bug button, then the server will gather what information it has about whether it has errors to send to a client, and what's blocking it from doing so. And it'll send all this information as part of the "rage" response to appear in the bug report.

How is this different from logging similar information via Hh_logger.log to the server log? With this, (1) we have the information in a more convenient place in the bug report, and (2) we can ask what is the *current* state of the server, rather than merely what was the state at the last time the server chose to report via Hh_logger.log

Reviewed By: kmeht

Differential Revision: D16400378

fbshipit-source-id: a7ea40383dad4246a3baa242b2ab0cc3ebbd9b22