Add test case demonstrating a difference between shallow decl and legacy decl
Summary:
This test case demonstrates a difference between the behavior of shallow decl and legacy decl which is not possible to reconcile (as far as I can tell) without making extreme changes to the implementation of shallow decl (which would likely have a substantial negative performance impact).

See the added test case for an example and explanation.

Reviewed By: ljw1004

Differential Revision: D16388119

fbshipit-source-id: 9d046336ed7384a202868b82aa7671d4b6e7da33