Add typecheck_like_types target
Summary:
This target turns on CI for all like types features, similar to how the new_inference tests were run. Initially, the normal `typecheck` target won't have like types turned on until most of the test differences are burned down, and then we'll make it the default.

A lot of test differences are positional changes, which I'll investigate in follow up diffs. As of now, we're at ~75% parity with the standard test suite.

Reviewed By: kmeht

Differential Revision: D15979574

fbshipit-source-id: ad3a4612f273b7669e18dcc8a7e2fee4e12c663d