Launder values guarding includes/requires in tests.
Summary:
There was a concern expressed on D16369348 that tests using if (true) to guard require_once could break in the future if the optimizer starts folding that sort of thing, and laundering the values was suggested.

This diff applies laundering to all the other tests where we use `if ({0,1,false,true})` to guard a require or include.

While we're here, also fix formatting issues and make tests more __EntryPoint-y where possible.

Reviewed By: ricklavoie, markw65

Differential Revision: D16402902

fbshipit-source-id: b5126fae4f9ee4480318dbb0f933610a68b4687d