Fix keywords.php FFP test
Summary:
D7357681 broke this test long time ago (which was manifested by [`.exp` file changing](https://our.intern.facebook.com/intern/phabricator/differential/changeset/?ref=372790846) from CST to `Failure: initial parsing exception`), but it was missed among all the other benign changes and a new `.exp` file was committed instead of fixing the bug.

The bug is related to smart constructors: if you `Make` a node using `parser`, you need to either
- later on consume this node, by `Making` it's parent using the same (or "descendant") `parser`
- abandon both node and parser (when backtracking)

Failing to do so can lead to corrupting smart constructor state (since `parser` state was already affected by construction of node).

The code in D7357681 ends up making `attribute_spec` node, but then abandoning it in last match case, without reverting the parser to the right version. The fix is some additional cloning/backtracking.

This covered other changes that this test needed (looks like we banned some keywords in the meantime). The fix is splitting the test in two - with valid and invalid keywords.

The Rust version of this check (`VerifySmartConstructors`) is still missing, which is why we needed `.rust_todo` that this diff also deletes - Rust was succeeding, while OCaml was crashing. Now they both succeed.

Reviewed By: shiqicao

Differential Revision: D16434481

fbshipit-source-id: a31d9e92b29b72107e15393a46d1e53287b37e37