HackEventLogger: server workers should initialize logging when sampling rate is 0
Summary:
Workers don't set up logging if the sampling ratio is 0, but that means that logging doesn't work in workers in general.
Introducing a worker-specific init in this change.

I moved the init start log entry out of logger init, as it seems like an odd place (init, in this case, refers to server init, not logger init).

Reviewed By: young-k

Differential Revision: D16400389

fbshipit-source-id: 01751714bcf28a65c11bd1bec4b02fbc118c9f7b