New inference: constraint-based async lambda
Summary:
Under the flag `--new-inference-lambda` Hack will check lambdas using constraint-based inference, generating fresh type variables for parameters and return type that aren't specified. Unfortunately, for async lambdas it leaves the return type as `Awaitable<_>` where `_` is the `Tany` type, instead of using `Awaitable<#1>` where `#1` is a fresh type variable.

Fix is straightforward. Also added a test that unfortunately has to make use of `hh_show` to demonstrate that `Tany` is not used.

Reviewed By: CatherineGasnier

Differential Revision: D16442654

fbshipit-source-id: 2cd95163782d04fd62c7cc22ae100bad51a72ee2