Simplify resolve_ty
Summary:
The "resolve_ty" function is slow because it calls `localize_with_self` for every result.

We can get mostly good enough results if we avoid this localize call, and the performance benefits are significant.

For a large result set using one commonly used class with 1,290 class elements, it takes > 4 seconds to process 1,290 results through resolve_ty.  After this change it takes 18 milliseconds.

What changes as a result of not localizing results? - Mostly, the impact is on generic functions.  The generic parameters and generic return values are no longer localized, which means that you can see generic types appear.

An example:

```
class C {
  public function test(): C::Z { ... }

  public function myfunc(): void {
    $x = new C();
    $x->AUTO332
```

When autocomplete runs on the the above snippet, C->test() is one of the possible results.  The function signature of C->test() appears as `test (function(): _)` if the ty is localized, and `test (function(): C::Z)` without localization.

I see no reason to strongly prefer one over the other.

Reviewed By: ljw1004

Differential Revision: D16442469

fbshipit-source-id: ee32e125aa7cbce5e8dcf81e34257f9b30e7b094