Add option to call layoutBlocks before estimating inlining vasm cost
Summary:
The inliner now uses different weights for blocks depending on whether they're
in main/cold/frozen in order to estimate the inlining cost. In this scenario, it
makes sense to run layoutBlocks() to repartition the blocks into hot/cold based
on profiling data.  This diff adds an option for this, but it leaves it off by
default for now.

Reviewed By: jano

Differential Revision: D16451977

fbshipit-source-id: 54fb2a343821e4d47f01c4648016350d2afa2751