Reset original file mode after gross XHP attribute hack
Summary:
We added this hack in 2014 to temporarily switch to partial mode for XHP attributes because they didn't support type params or something (D1681272).

But, we never switched it back, which was okay, because we discarded the env. But over time, refactorings let the env escape that expression, and finally in D15413998, it escaped the function—meaning that any file containing XHP with attributes were being checked in partial mode after the first attribute.

This is a quick fix to restore previous behavior, while the followup diff removes the behavior entirely. We should probably kill `Typing_env.set_mode` with fire.

Reviewed By: DavidSnider

Differential Revision: D16455881

fbshipit-source-id: d390b57553cc3f8e1a556874e2e8cf78e4d34a12