Added hh_rewrite_declarations binary
Summary:
Added hh_rewrite_declarations binary, its [logic](https://fb.quip.com/A6bDAAHD5Y01) and some tests. The purpose of this is to be used in tandem with the hack kernel for Bento (D16417320).
This approach will probably be modified soon, such that fb_intercept will be used instead!

Reviewed By: arun11111

Differential Revision: D16438624

fbshipit-source-id: a7a607725b407db7c2335e5df63863833dad3da6