Generate function declaration based on Typing_defs.fun_type
Summary: When a function which we want to extract calls another function, we need to get the latter function's declaration. We used to just read it from the file where the function is declared, which is not efficient. This commit introduces generating the declaration based on Typing_defs.fun_type.

Reviewed By: andrewjkennedy

Differential Revision: D16261224

fbshipit-source-id: b68b1eb3aef6a4326634876ddbba185b1d1e563b