Support self-referential DefLabels in pdce-inlining
Summary:
Originally I thought that we shouldn't get any
self-referential DefLabels containing FramePtrs, so any instance of
them would be a bug and we should assert. I now see how they can arise
because of loops, so modify pdce-inlining to handle them. We should
just ignore the self-referential portion of the DefLabel (there should
always be a portion which isn't).

Reviewed By: markw65

Differential Revision: D16456697

fbshipit-source-id: 86e80fbf71e3984c788ef0c3c54b4b19552e6593