Drop `telemetry/event` notifications in testing
Summary: These seem a little bit non-deterministic. Although there appear to be generated deterministically, they're noisy and the rules for when they're generated aren't immediately obvious. In the previous commit, we add the `powered_by` field to JSONRPC responses and notifications, which is a much more direct and reliable way of determining which back-end served a request, rather than relying on the presence of these `telemetry/event` messages.

Reviewed By: jewelpit

Differential Revision: D16370667

fbshipit-source-id: 81dbaa3933e636d3703028af72198c9ff60a2bfc