Wait for `shutdown` response in LSP tests
Summary: Rather than read a certain number of expected requests from the language server in tests, simply wait for the shutdown message to be sent. This is useful in future commits where we add more meta-language commands (such as `$test/waitForResponse`) which consume responses, and therefore make the response count not a useful metric.

Reviewed By: 2BitSalute

Differential Revision: D16442120

fbshipit-source-id: a3473c885c46d0106e483c868d8b7ef03b4785ef