Only cache resolve() results in pdce-inlining
Summary:
Due to the changes to allow for self-referential DefLabels in
pdce-inlining, the results of resolveDefLabel() are now path dependent
(that is, it depends on the order you visit instructions). Because of
this, its not correct to cache each intermediate lookup. Instead only
cache the final result (the one returned to resolve()).

Reviewed By: jano

Differential Revision: D16475387

fbshipit-source-id: 6eb5ba4064b1a80fdf02f565b85fd1a5bca03422