Add visibility check of private/protected member for constraints
Summary:
[hack] Typing_visibility now considers subtyping with where constraints

This diff enables visibility checking when a class with constraints tries to access members belonging to constraints. This feature is important for constraints to serve as "required classes" rather than merely upper bounds in subtyping.

Visibility check can be performed via subtyping:
1. public visibility means no checks
2. protected visibility means the caller is a subtype of the class where the method was declared.
3. private visibility means the caller and the class where the method was declared are mutually subtype of the other.

With this diff, we will be able to write partial classes in Hack with the constraint `trait PartialC where this = C`.

Reviewed By: Matt-Schellhas

Differential Revision: D16348409

fbshipit-source-id: bec35e23a86a3bf0dbc8f583e9f1d20b9f4eb16b