Include name of inheriting class in Decl_heap_elems_bug message
Summary:
While investigating T47852226, it was useful for me to know which class inherited the missing member (because that class was an invalid declaration stored in the saved state). This logging seems useful in general and I think we should keep it.

This change also avoids running the `sprintf` unless we are actually going to raise the exception.

Reviewed By: 2BitSalute

Differential Revision: D16495861

fbshipit-source-id: ca924a375d01722c18e8c41fe439c36e59c6fe5e