Do not needlessly compute type of result of array assignment
Summary: The code in `Typing_array_access` for dealing with array assignments of the form `$x[$y] = $z` computes the TAST type of the whole assignment expression recursively. But as far as I can tell, this will simply produce the type of `$z`, which we have in our hand already. In any case, for inference purposes we use the type of `$z` as the type (see line in `typing.ml`), so for consistency we should just use this in the TAST also.

Reviewed By: kmeht

Differential Revision: D16494603

fbshipit-source-id: d59d6ef6c198912376980566ce08545d8a16ddcb