Remove class-refs
Summary:
Class-refs were added as a way to avoid needing a KindOfClass
datatype. Before that, KindOfClass values could exist on the stack
(but nowhere else), and weren't a subtype of TGen in the JIT. Since
Class* was only stored in a class-ref, we never needed a
datatype. However, we're not in a space crunch for datatypes anymore,
and KindOfClass is back anyways. So, there's no longer much of a
compelling reason to have them.

This diff removes the usage of class-ref slots for all
bytecodes. Instead of writing or reading a class-ref slot, the
instruction pushes or pops a Class* to the stack, like other
values. This is actually less complicated than the original way, since
TCls are a subtype of TGen now, so there's very little special
handling needed.

The only real complication is that the class-ref slots grew to contain
reified generic information as well. However, an audit of the emitter
reveals that there's actually only one combination of bytecodes where
the reified generic information might be emitted and might be
used. That is: ClsRefGetTS and NewObj[D]. ClsRefGetTS is renamed to
ClassGetTS and pushes two values on the stack: The Class and the
reified generics. We use KindOfNull to represent the lack of reified
generic information (since we can't use a bare null pointer on the
stack like previously). Add a new bytecode called NewObjR which
behaves like NewObj, but expects reified generic information to be on
the stack as well. NewObj now is just for the case where there's no
reified generic information (it just pops the Class). These situations
can be distinguished statically in the emitter (this also means the
"MaybeReified" immediate stuff isn't needed anymore).

BaseSC has a few complications as well. Due to order of operation
issues, it can read its property name at an arbitrary offset into the
stack. Likewise, the Class needs to be evaluated (and hence pushed
onto the stack) at different points. Therefore the bytecode cannot
just pop the Class off the top. Add a new immediate to BaseSC which
gives the offset into the stack of the Class as well. This required
adding some additional support to HackC to track the different
possible locations of the property name and class.

Other than that, everything is straightforward. DiscardClsRef is gone,
it can be done with PopC now. Rename ClsRefGetC and ClsRefGetTS to
ClassGetC and ClassGetTS. Rename ClsRefName to ClassName. Most
bytecode changes (like for Self, etc) are straightforward.

Note that this diff leaves a *lot* of dead code. For easier review,
that will all be removed in a follow-up diff.

Reviewed By: jano, markw65

Differential Revision: D16374292

fbshipit-source-id: 00a11023187ad04767d1c7014ac0c35e6e08cc7a