Convert WatchmanAutoloadMap::sync to a Future
Summary:
(Most of this diff is closed-source, but all the code will be open-sourced as soon as we get our own house in order.)

This diff adds an asynchronous implementation of `WatchmanAutoloadMap::sync`, the function which synchronizes the WatchmanAutoloadMap with the repo's state. I didn't like having a method called `sync` which is implemented in an async manner, so I renamed the async method to `update` and the synchronous version to `ensureUpdated`. In Facebook Hacklang code we usually prefix async methods with `gen`, e.g. `async function genUpdate(): Awaitable<bool>`, but I don't know if we have any such convention for C++ methods using folly Futures.

Note that I'm still using a lock here, and I'm still staying in the lock while I wait for the future chain to complete. This is because I'm moving futures into each other to set up a future chain, and the move operation is not thread-safe. We only ever have one update operation going on at a time anyway, so I don't think the lock synchronizes our resources beyond what the problem domain requires.

Reviewed By: dneiter

Differential Revision: D16424845

fbshipit-source-id: 0dd60bc38d4066efb954eaf406853d3b51ae86fc