De-duplicate Class code for handling redeclared instance props.
Summary: De-duplicate the code that handles redeclaring instance properties in Class::setProperties. The case where we were redeclaring with a public property VS with a protected property differed only in the presence/absence of a 2-line if that weakens a protected property to public. Move the code into a lambda that works for both cases.

Reviewed By: paulbiss

Differential Revision: D16434162

fbshipit-source-id: 219991dd2b0c40542630f899e97a518c95712499