Add __NoFlatten to hhbbc, use it to fix const object test.
Summary:
Add an attibute that tells hhbbc to not flatten traits marked with it. This lets us reliably write unit tests for non-flattened traits without jumping through hoops.

Use it to fix hphp/test/slow/object/const/copied-from-non-flattened-trait.php. Since it was written, a change in hhbbc caused traits to get flattened in it despite the hack it was trying to use to prevent that, and another change caused us to turn NewObjS into NewObjD, which meant we would resolve class C by name in D::makeParent, think it was non-unique, and refuse to optimize the LockObj away without testing the code paths that this test is trying to cover. By forcing T to not be flattened into C without using hacks, and by making C unique, we avoid both of these issues and the test once again tests what we want.

Reviewed By: oulgen

Differential Revision: D16434160

fbshipit-source-id: c3b8df12f1b20f98bf9c9b7fa66e4fa416778c61