Pass through RPC commands
Summary:
No behavioral change. There was originally a layer of abstraction in `ClientIdeService`, where we exposed each RPC command as an individual function. I thought this might be useful for integration tests or something, but it has turned out to be unnecessary altogether. Remove it, and instead have the callers make direct calls to `rpc`.

This is also useful for the future where we want to pass a "notification handler" to `rpc`, so that we can handle notifications that come in while we're waiting for a certain RPC request to complete. We would have to add this notification handler as a parameter to every IDE command.

This commit also unifies the concept of a "document location", which identifies the cursor at a certain location in the document. This is one of the most common parameters to an IDE command.

Differential Revision: D16465331

fbshipit-source-id: 8ccd1d164defa46d12c7a4750854bec992b775a1