Rename constructNoConst to lockWhileUnwinding, move the default from false to true
Summary: In order to support the JIT unwinder, we need to be able to make decisions on whether to lock the object at the top without having to rely the particular bytecode. Changing the FCA in this way so that `lockWhileUnwinding` is set only when we need to lock the object at the top enables us to emit a more generic codepath.

Reviewed By: alexeyt

Differential Revision: D16522474

fbshipit-source-id: 719e854c66652fd57bba4e1fea5c997a4be98dc5