Simplify class_get implementation (part 1)
Summary:
The `class_get` function in `typing.ml` is unnecessarily complex, and without refactoring isn't amenable to using D16498059 for the intersection, union, and upper bounds cases.

This is a first step at simplification: observing that the `Errors.abstract_const_usage` error is the only client of the `cc_abstract_info` result, and so we can avoid passing back this information in the result tuple, and instead register the error when picking up the information.

Reviewed By: kmeht

Differential Revision: D16506182

fbshipit-source-id: eee954d1b07ed1eaed094a2b09a07ecbfac84c2b