client rage only gives server 30s to produce its rage report
Summary:
We've had a spate of bug reports that just show "TIMED OUT".

The probable cause is that pstack just took too long.

But I'm also concerned that clientLsp might fail to deliver rage reports in a timely fashion in cases where the server is unresponsive.

This diff introduces a timeout of 30s for the server to respond with rage results.

Reviewed By: jewelpit

Differential Revision: D16553014

fbshipit-source-id: e5a5a5071b6c0cc810fc41c1a3996a81c7b1be99