Ban classname/typename as reified generics (top-level)
Summary: This bans classname and typename from being reified at the top level. But, this doesn't drill through type aliases—for that, it will have to be moved to use the same type validator pattern as is/as operators and disallowing PHP arrays in type constants. Worth noting that the type alias bug is true for all types that we are supposed to ban, so fixing this is required to ship the feature.

Reviewed By: oulgen

Differential Revision: D16557181

fbshipit-source-id: 54ef8ca72c8b35296a06e00d6f6bb1352f646979