Implement generic typing rules for unions, intersections, and upper bounds on abstract types
Summary:
Currently we re-implement the same logic again and again for propagating unions, "trying" all intersections, and delegating to upper bounds of generic parameters and abstract types. This is done for array get, array append, array assignment, method invocation, property access, and some other constructs.

I would like to factor this out by introducing a new function `Typing_generic_rules.apply_rules env ty f` that uses the function `f` to apply a typing rule specific to a particular construct on a concrete type (e.g. for method invocation: a classish type; for array access: a collection type) but deal with unions, intersections, and bounds on abstract types in a single piece of code.

In this diff, the generic code is introduced, and applied to array operations. Subsequent diffs will deal with method invocation and other constructs.

Reviewed By: vassilmladenov

Differential Revision: D16498059

fbshipit-source-id: 6c5776d8a1516e1fd63dd40b7979e385b7f47263