Convert Tany immediately following Errors into Terr
Summary:
`Tany` is inherently unsound, behaving as the sub and super-type of all other types.

This converts a few simple situations into `Terr`, which has the same semantics but will require someone to suppress an error to get the broken behavior.

Also, it uncovered a obscure bug where we allowed a `Tany` inout parameter, but not a `Terr`. This was "fixed" in `callconv_check.ml`.

Reviewed By: vassilmladenov

Differential Revision: D16550282

fbshipit-source-id: 0f1843f8d6e4ff1c17dea786ac7e767990ce7a61