Add Type_validator module
Summary:
This refactors duplicated logic across four modules to create `type_validator`, which extends `type_visitor` with validation logic for both decl and locl types. The functionality was originally added to validate the hints used for `is` expressions, but has become useful elsewhere in our typechecking as well.

As a proof of concept, Typing_coercion has been updated to use the new validator. Followup diffs will convert the other usages.

Reviewed By: Matt-Schellhas

Differential Revision: D16557900

fbshipit-source-id: f9fd8a5297fdc0f8080ca7b3d656e0f1c930cc93