Inline and simplify JIT class method loading helpers
Summary:
All but one caller of `fpushClsMethodKnown()` and `lookupClsMethodKnown()`
pass a set of booleans that cause these methods to become trivial.

Inline these methods to their callers and simplify the resulting code.

Allows `fpushClsMethodD()` to use `prepareToCallKnown()`, brings the
implementation of `FPushClsMethod*` opcodes closer to `FCallClsMethod*` and
makes it slightly easier to replace `NewClsMeth` with `cns(...)` in LOWPTR
mode.

Reviewed By: ricklavoie

Differential Revision: D16553920

fbshipit-source-id: fa4cbb62d026873c1f0eeda1812ab2f4422077e6