Fix positioning of aliased namespaces
Summary:
The function `get_replace_pos_exn` does not correctly calculate the replacement position for text.  When the user uses an aliased namespace such as `Str\compare` (which is actually `HH\Lib\Str\compare`), the function gets lost: it attempts to fast forward the start position all the way through the fully un-aliased namespace, while it attempts to move the ending position backwards by the distance of the AUTO332 text.  The end result is that the end position often comes before the start position.

Yuck!

Let's simplify the logic as follows:

* We know the correct "end" position: that's the position where the text AUTO332 existed.
* We know that the query text shows up as "HH\Lib\Str\cAUTO332".
* We know that we will always delimit on namespaces, so everything before the final backslash should be ignored.

Based on that, the correct calculation is:
* Take the end position exactly as-is.
* The start position should be backspaced by the length of the name once you have eliminated both AUTO332 and the namespace.

Reviewed By: 2BitSalute

Differential Revision: D16551192

fbshipit-source-id: 4ec02a5688f0d1691d6f9e1937213c9800185625