send SUBSCRIBE_DIAGNOSTIC asap
Summary: I'm still trying to track down the issue where Nuclide isn't getting diagnostics from hh. It looks like the server doesn't believe a client is subscribed. That should be impossible. Nevertheless, I'm moving the subscribe message up a notch, in case there was something in the client or server handling of open-file-events that was breaking the diagnostic subscription.

Reviewed By: 2BitSalute

Differential Revision: D16567718

fbshipit-source-id: 4d7f363056f478b212085d4e840fdbddeb7b7662