Simplify obj_get implementation (part 1)
Summary:
The implementation of `obj_get` is needlessly complex. Rather than plumb around visibility information, to be used *only* for `inst_meth` checks upon return, let's check the visibility at the point that it is obtained.

I believe this subsumes the use of `min_vis`, as the use of the most restrictive visibility just falls out from repeated checks.

Reviewed By: vassilmladenov

Differential Revision: D16560932

fbshipit-source-id: 0442b4a13f95dfa771be1c8b2a8f71ee22af8c32