Add inlining support for inout functions
Summary:
We depend on inlining to reduce the overhead of builtin calls in a lot of instances. As the next diff in this stack will add support for inout builtins this diff is needed to ensure they're at parity with reffy builtins.

Also fixes some bugs with the way return types for inout functions are decoded and adds some type inference for out parameters on function return.

Reviewed By: ricklavoie

Differential Revision: D16525909

fbshipit-source-id: 0c4cbfcadb4fd39c7329a843274d2aa07d48f185