FCallBuiltin support for inout
Summary: Allows hhbbc to optimize inout native functions to FCallBuiltin. This required new push/pop flavors for FCallBuiltin allowing it to pop Uninit values and push out values to the stack, and a new immediate indicating the number of out values. The PopFrame bytecode was also added to allow the Uninit frame locations to be popped from the stack following the FCallBuiltin. In most cases it will be optimized away by hhbbc's DCE pass.

Reviewed By: ricklavoie

Differential Revision: D16544757

fbshipit-source-id: 1d6441fe1ec476ea955074fcb04310109cb008b3