FPushClsMethod*: write ActRec fully in JIT
Summary:
Do not write function pointer and class context or `$this` in LookupClsMethod.
Instead, return the function pointer and use `prepareToCallUnknown()` standard
mechanism to write ActRec.

Previously, LookupClsMethod was responsible for forwarding `$this` if the
target method was non-static and it was called from an instance method in the
same class hierarchy. This behavior will be removed and this diff handles it by
returning nullptr from LookupClsMethod and exiting to the interpreter.

Reviewed By: ricklavoie

Differential Revision: D16572538

fbshipit-source-id: 2a3d80183922e0fbc81cdaf84dd01ffef0a30617