Remove most usage of currentNormalizedInstruction
Summary:
currentNormalizedInstruction is not guaranteed to be up to date in the middle
of an opcode (e.g. because we inlined other stuff, which overwrote the state).

Instead of fixing it, access the same data via curSrcKey(env), which is
properly maintained to be up to date.

Allows the next diff to use makeExitSlow() in a slow path after fast path that
possibly involved inlining was emitted.

The last remaining usage is for accessing funcd, that one will be killed
together with legacy FCall opcode in the future diff.

Reviewed By: ricklavoie

Differential Revision: D16577309

fbshipit-source-id: bcaf267a377f2798404bf16e2ed654ada26c6ca8