Forbid goto labels when the goto operator is forbidden
Summary:
The goto operator is supported in hackc and the runtime because it is used by the output of the coroutine rewriter, but it is not supported in Hack files, and not understood by the typechecker. We have an error for uses of the goto operator in Hack files, but not for goto labels. These labels are not harmless--the added test case stray_goto_label.php should emit an error about how `int` is not compatible with `string`, but it doesn't, because labelled statements are not properly understood by the typechecker.

Labels are used only for goto (and not `break`), so when the goto operator is forbidden, it does not make sense to permit a goto label, and we can simply emit an error.

Reviewed By: hgoldstein

Differential Revision: D16581188

fbshipit-source-id: c41ce907dae081cbe866b101b7478ba5f2aef845