Remove obsolete test
Summary:
I've removed this test because it wasn't testing anything useful, and the output will change when we remove Aast.Any.

It was added in D6567018 to check that we didn't allow inout parameters with variable variables. We then banned variable variables in D6705828 and just FIXME'd it here.

Since variable variables are gone, this test isn't exercising anything.

Reviewed By: kmeht

Differential Revision: D16581586

fbshipit-source-id: eb2e069a5b08bd1574f34b79b96babece1327f05