hh --pause: implementation
Summary:
This diff fulfills the vision that `hh --pause` will make it so that file changes no longer trigger a full typecheck.

It's surprisingly easy because hh_server already keeps track of everything it needs - namely, (1) a list of all disk-files that have changed and will need to be checked, (2) flags `Full_check_needed` and `Full_check_started`

The existing meaning of `Full_check_needed` is that some disk changes have not yet been fully processed. The existing way we get into this state comes from D7944023, when a full-type-check was underway but then an IDE operation happened which interrupted the type check. So it's an existing legitimate state that's already handled correctly. That's why I think it's fine to use for `hh --pause`

Incidentally, because `hh --pause` is triggered by a command after initialization has completed, it's kind of nice because I didn't have to worry about whether pause mode would interact badly with interaction.

Note that IDE error squiggles continue to work fine. They only ever used local checks, which are unaffected by this diff.

Although I think the feature works, and seems correct, and passed my tests, I've left `hh --pause` documented as EXPERIMENTAL. My thought is that the primary audience right now is us as we develop hulk, and after that maybe only the first adopters of hulk v1. This feature will become irrelevant once we move to the new architecture.

Reviewed By: 2BitSalute

Differential Revision: D16574011

fbshipit-source-id: 08892d74c5cda60f01f341486f6d8a84a92af05a