hh --pause: plumbing
Summary:
Currently, hh_server does a full recheck whenever it discovers a file has been changed on disk.

This is a pain in a bunch of scenarios! e.g. I want to make a load of changes, but defer the full recheck until I'm ready.

My end goal is that users can write `hh --pause` to turn off that "trigger full recheck upon file change". Some users will prefer to live permanently in `hh --pause` mode, and the only time a full check happens is when they explicitly trigger it with `hh check`. Other useres will switch out of the mode with `hh --resume`.

This diff is solely the dull plumbing to support `hh --pause` and `hh --resume`. It ultimately winds up in a variable inside ServerEnv. This diff doesn't yet use that variable (that will be left to the next diff in the stack).

Why did I implement this as something the user turns on and off (rather than via hh.conf)? Because
1. Users don't have easy ability to change their own hh.conf
2. Changing hh.conf is a pain because it requires a restart
3. I wanted to support the scenario where users dynamically chose to enter or leave `hh --pause`

In the future once we move to VSCode, then this will become irrelevant. We'll be doing the watching then via a standalone monitor process. Folks who want `hh --pause` will simply not launch (or will terminate) the monitor. Folks who want `hh --resume` will relaunch the monitor.

Reviewed By: 2BitSalute

Differential Revision: D16573714

fbshipit-source-id: abcd9dd46c17eaf9abc7f11d2b53dce2b851e6ac