workaround for MacOS hh_client hang (#8546)
Summary:
This change introduces a workaround for issues raised in https://github.com/facebook/hhvm/issues/8322.

On MacOS, hh_client will hang in situations where the server process takes a long time to get to its "hello" response. The underlying cause appears to be that fds sent via Unix domain sockets can go stale on the receiving side if the equivalent fds are closed from the sending process. 72ab5dd addressed the same issue, but works only when server startup time is relatively fast.

One workaround for longer startup times is having the client re-establish a connection to the monitor in cases where it took a long time to get to the server hello. This is a little clumsy but better than the alternative, which is to manually kill hh_client. A true fix would require a bit more research, and may be moot given that [there are plans](https://github.com/facebook/hhvm/issues/8322#issuecomment-515128629) to re-architect hh_client/hh_server IPC.
Pull Request resolved: https://github.com/facebook/hhvm/pull/8546

Differential Revision: D16565466

fbshipit-source-id: ec028579b07ff0a2550b5842a80a9e282989ad58