Simplify the variadic_hint type in AST/AAST
Summary:
Now that variadic function hints must have a type, we can simplify the structure and turn it into an option. `None` indicates the function isn't variadic, `Some hint` indicates it is, with the given hint.

This will make it easier to implement variadic function hints in TypeStructure, which in turn will allow those hints to be reified type arguments.

Reviewed By: oulgen

Differential Revision: D16420665

fbshipit-source-id: 89979cea930c3f6902b73b610a5a81c813c106d2