Fix raise_notice() usage in unserialize_ex()
Summary:
If unserialization failed, we'd attempt to print out the bad
string (but only up to the first 1000 characters). The problem is,
there's no guarantee the string is null terminated (we have an
explicit length field). So, vsnprintf would potentially read data past
the end of the string, until it hits a null terminator, or hits 1000
characters. Fix this by cap the maximum number of printed characters
to the minimum of the given length or 1000.

Reviewed By: alexeyt

Differential Revision: D16586657

fbshipit-source-id: 74fe529b9d93d006f4a1504b82cde34c7b33e8b4