Updating HH\Pair documentation to reflect it's immutability
Summary:
HH\Pair is actually immutable, and is backed by an Immutable Vector.  A Pair can't be created empty, and attempting to set the individual members (indices 0 and 1) fails due to the underlying storage being immutable as well.  This also reflects the [language in the `immutable` function](https://fburl.com/m5no57yd) which doubly confirms that Pairs are, in fact, immutable.

Also fixing a typo or two.

Reviewed By: nilesr, hgoldstein

Differential Revision: D16583235

fbshipit-source-id: cef8dd141551ceebed77fa9ac24e226244855fc2